requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# ALPR Engine - Dépendances Python (LEAD Production)
# ═══════════════════════════════════════════════════════════════════════════════

# ───────────────────────────────────────────────────────────────────────────────
# CORE - Interface & Framework
# ───────────────────────────────────────────────────────────────────────────────
gradio>=4.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# ───────────────────────────────────────────────────────────────────────────────
# ML - Détection & Classification
# ───────────────────────────────────────────────────────────────────────────────
ultralytics>=8.0.0
fast-plate-ocr
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0              # EfficientNet et autres architectures
onnxruntime>=1.16.0

# ───────────────────────────────────────────────────────────────────────────────
# IMAGE & VIDEO
# ───────────────────────────────────────────────────────────────────────────────
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# ───────────────────────────────────────────────────────────────────────────────
# DATABASE
# ───────────────────────────────────────────────────────────────────────────────
# SQLite (inclus dans Python)
psycopg2-binary>=2.9.9   # PostgreSQL (Neon DB)
sqlalchemy>=2.0.0        # ORM optionnel

# ───────────────────────────────────────────────────────────────────────────────
# STOCKAGE S3
# ───────────────────────────────────────────────────────────────────────────────
boto3>=1.34.0            # AWS S3 / Scaleway / MinIO
botocore>=1.34.0

# ───────────────────────────────────────────────────────────────────────────────
# MLOPS - Tracking & Registry
# ───────────────────────────────────────────────────────────────────────────────
mlflow>=2.10.0

# ───────────────────────────────────────────────────────────────────────────────
# SERVING - Ray Serve (optionnel en dev)
# ───────────────────────────────────────────────────────────────────────────────
ray[serve]>=2.9.0

# ───────────────────────────────────────────────────────────────────────────────
# SCHEDULING - Airflow (optionnel, installé séparément en prod)
# ───────────────────────────────────────────────────────────────────────────────
# apache-airflow>=2.8.0  # Installer séparément si nécessaire

# ───────────────────────────────────────────────────────────────────────────────
# UTILS
# ───────────────────────────────────────────────────────────────────────────────
pandas>=2.0.0
python-dotenv>=1.0.0     # Chargement .env (backup)
httpx>=0.26.0            # HTTP client async
tenacity>=8.2.0          # Retry logic
pyyaml>=6.0.0            # Config YAML

# ───────────────────────────────────────────────────────────────────────────────
# DEV & TEST (optionnel)
# ───────────────────────────────────────────────────────────────────────────────
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.1.0
